Moved contact info parts to new classes and added update endpoints
