+++ rest/src/main/java/fi/otavanopisto/pyramus/rest/StudentRESTService.java
@@ -37,8 +37,6 @@ import org.apache.commons.lang3.StringUtils;
 
 import fi.otavanopisto.pyramus.dao.base.OrganizationDAO;
 import fi.otavanopisto.pyramus.domainmodel.Archived;
-import fi.otavanopisto.pyramus.domainmodel.base.Address;
-import fi.otavanopisto.pyramus.domainmodel.base.ContactType;
 import fi.otavanopisto.pyramus.domainmodel.base.ContactURL;
 import fi.otavanopisto.pyramus.domainmodel.base.ContactURLType;
 import fi.otavanopisto.pyramus.domainmodel.base.Curriculum;
@@ -50,7 +48,6 @@ import fi.otavanopisto.pyramus.domainmodel.base.Municipality;
 import fi.otavanopisto.pyramus.domainmodel.base.Nationality;
 import fi.otavanopisto.pyramus.domainmodel.base.Organization;
 import fi.otavanopisto.pyramus.domainmodel.base.Person;
-import fi.otavanopisto.pyramus.domainmodel.base.PhoneNumber;
 import fi.otavanopisto.pyramus.domainmodel.base.School;
 import fi.otavanopisto.pyramus.domainmodel.base.StudyProgramme;
 import fi.otavanopisto.pyramus.domainmodel.base.StudyProgrammeCategory;
@@ -76,7 +73,6 @@ import fi.otavanopisto.pyramus.domainmodel.users.StaffMember;
 import fi.otavanopisto.pyramus.domainmodel.users.User;
 import fi.otavanopisto.pyramus.domainmodel.users.UserVariable;
 import fi.otavanopisto.pyramus.domainmodel.users.UserVariableKey;
-import fi.otavanopisto.pyramus.framework.UserEmailInUseException;
 import fi.otavanopisto.pyramus.framework.UserUtils;
 import fi.otavanopisto.pyramus.rest.annotation.RESTPermit;
 import fi.otavanopisto.pyramus.rest.annotation.RESTPermit.Handling;
@@ -2591,64 +2587,6 @@ public class StudentRESTService extends AbstractRESTService {
     return Response.noContent().build();
   }
 
-  @Path("/students/{STUDENTID:[0-9]*}/emails")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response listStudentEmails(@PathParam("STUDENTID") Long studentId) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.LIST_STUDENTEMAILS }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    List<Email> emails = student.getContactInfo().getEmails();
-    if (emails.isEmpty()) {
-      return Response.noContent().build();
-    }
-
-    return Response.ok(objectFactory.createModel(emails)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/emails")
-  @POST
-  @RESTPermit(handling = Handling.INLINE)
-  public Response createStudentEmail(@PathParam("STUDENTID") Long studentId, fi.otavanopisto.pyramus.rest.model.Email email) {
-    if (email == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.CREATE_STUDENTEMAIL, StudentPermissions.STUDENT_OWNER }, student, Style.OR)) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    Long contactTypeId = email.getContactTypeId();
-    Boolean defaultAddress = email.getDefaultAddress();
-    String address = email.getAddress();
-
-    if (contactTypeId == null || defaultAddress == null || StringUtils.isBlank(address)) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    ContactType contactType = commonController.findContactTypeById(contactTypeId);
-    if (contactType == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    try {
-      return Response.ok(objectFactory.createModel(studentController.addStudentEmail(student, contactType, address, defaultAddress))).build();
-    } catch (UserEmailInUseException ueiue) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-  }
-  
   @Path("/students/{STUDENTID:[0-9]*}/defaultemail")
   @GET
   @RESTPermit(handling = Handling.INLINE)
@@ -2668,311 +2606,6 @@ public class StudentRESTService extends AbstractRESTService {
     }
 
     return Response.ok(objectFactory.createModel(email)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/emails/{ID:[0-9]*}")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response findStudentEmail(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.FIND_STUDENTEMAIL }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    Email email = commonController.findEmailById(id);
-    if (email == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!email.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    return Response.ok(objectFactory.createModel(email)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/emails/{ID:[0-9]*}")
-  @DELETE
-  @RESTPermit(StudentPermissions.DELETE_STUDENTEMAIL)
-  public Response deleteStudentEmail(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    Email email = commonController.findEmailById(id);
-    if (email == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!email.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    commonController.deleteEmail(email);
-
-    return Response.noContent().build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/addresses")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response listStudentAddresses(@PathParam("STUDENTID") Long studentId) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.LIST_STUDENTADDRESSS }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    List<Address> addresses = student.getContactInfo().getAddresses();
-    if (addresses.isEmpty()) {
-      return Response.noContent().build();
-    }
-
-    return Response.ok(objectFactory.createModel(addresses)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/addresses")
-  @POST
-  @RESTPermit(handling = Handling.INLINE)
-  public Response createStudentAddress(@PathParam("STUDENTID") Long studentId, fi.otavanopisto.pyramus.rest.model.Address address) {
-    if (address == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.UPDATE_STUDENT, StudentPermissions.STUDENT_OWNER }, student, Style.OR)) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    Long contactTypeId = address.getContactTypeId();
-    Boolean defaultAddress = address.getDefaultAddress();
-    String name = address.getName();
-    String streetAddress = address.getStreetAddress();
-    String postalCode = address.getPostalCode();
-    String country = address.getCountry();
-    String city = address.getCity();
-
-    if (contactTypeId == null || defaultAddress == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    ContactType contactType = commonController.findContactTypeById(contactTypeId);
-    if (contactType == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    return Response.ok(
-        objectFactory.createModel(studentController.addStudentAddress(student, contactType, defaultAddress, name, streetAddress, postalCode, city, country)))
-        .build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/addresses/{ID:[0-9]*}")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response findStudentAddress(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.FIND_STUDENTADDRESS }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    Address address = commonController.findAddressById(id);
-    if (address == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!address.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    return Response.ok(objectFactory.createModel(address)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/addresses/{ID:[0-9]*}")
-  @PUT
-  @RESTPermit(handling = Handling.INLINE)
-  // @RESTPermit (StudentPermissions.UPDATE_STUDENTADDRESS)
-  public Response updateStudentAddress(
-      @PathParam("STUDENTID") Long studentId,
-      @PathParam("ID") Long id,
-      fi.otavanopisto.pyramus.rest.model.Address body
-      ) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(
-          new String[] { StudentPermissions.UPDATE_STUDENTADDRESS },
-          student)
-        && !restSecurity.hasPermission(
-            new String[] { StudentPermissions.STUDENT_OWNER },
-            student )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    Address address = commonController.findAddressById(id);
-    if (address == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!address.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-    
-    ContactType contactType = commonController.findContactTypeById(body.getContactTypeId());
-    
-    address = studentController.updateStudentAddress(
-        address,
-        contactType,
-        body.getDefaultAddress(),
-        body.getName(),
-        body.getStreetAddress(),
-        body.getPostalCode(),
-        body.getCity(),
-        body.getCountry());
-
-    return Response.ok(objectFactory.createModel(address)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/addresses/{ID:[0-9]*}")
-  @DELETE
-  @RESTPermit(StudentPermissions.DELETE_STUDENTADDRESS)
-  public Response deleteStudentAddress(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    Address address = commonController.findAddressById(id);
-    if (address == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!address.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    commonController.deleteAddress(address);
-
-    return Response.noContent().build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/phoneNumbers")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response listStudentPhoneNumbers(@PathParam("STUDENTID") Long studentId) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.LIST_STUDENTPHONENUMBERS }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    List<PhoneNumber> phoneNumbers = student.getContactInfo().getPhoneNumbers();
-    if (phoneNumbers.isEmpty()) {
-      return Response.noContent().build();
-    }
-
-    return Response.ok(objectFactory.createModel(phoneNumbers)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/phoneNumbers")
-  @POST
-  @RESTPermit(handling = Handling.INLINE)
-  // @RESTPermit (StudentPermissions.CREATE_STUDENTPHONENUMBER)
-  public Response createStudentPhoneNumber(@PathParam("STUDENTID") Long studentId, fi.otavanopisto.pyramus.rest.model.PhoneNumber phoneNumber) {
-    if (phoneNumber == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.UPDATE_STUDENT, StudentPermissions.STUDENT_OWNER }, student, Style.OR)) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-    Long contactTypeId = phoneNumber.getContactTypeId();
-    Boolean defaultNumber = phoneNumber.getDefaultNumber();
-    String number = phoneNumber.getNumber();
-
-    if (contactTypeId == null || defaultNumber == null || StringUtils.isBlank(number)) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    ContactType contactType = commonController.findContactTypeById(contactTypeId);
-    if (contactType == null) {
-      return Response.status(Status.BAD_REQUEST).build();
-    }
-
-    return Response.ok(objectFactory.createModel(studentController.addStudentPhoneNumber(student, contactType, number, defaultNumber))).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/phoneNumbers/{ID:[0-9]*}")
-  @GET
-  @RESTPermit(handling = Handling.INLINE)
-  public Response findStudentPhoneNumber(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    if (!restSecurity.hasPermission(new String[] { StudentPermissions.FIND_STUDENTPHONENUMBER }, student) && !restSecurity.hasPermission(new String[] { PersonPermissions.PERSON_OWNER }, student.getPerson() )) {
-      return Response.status(Status.FORBIDDEN).build();
-    }
-
-    PhoneNumber phoneNumber = commonController.findPhoneNumberById(id);
-    if (phoneNumber == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!phoneNumber.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    return Response.ok(objectFactory.createModel(phoneNumber)).build();
-  }
-
-  @Path("/students/{STUDENTID:[0-9]*}/phoneNumbers/{ID:[0-9]*}")
-  @DELETE
-  @RESTPermit(StudentPermissions.DELETE_STUDENTPHONENUMBER)
-  public Response deleteStudentPhoneNumber(@PathParam("STUDENTID") Long studentId, @PathParam("ID") Long id) {
-    Student student = studentController.findStudentById(studentId);
-    Status studentStatus = checkStudent(student);
-    if (studentStatus != Status.OK)
-      return Response.status(studentStatus).build();
-
-    PhoneNumber phoneNumber = commonController.findPhoneNumberById(id);
-    if (phoneNumber == null) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    if (!phoneNumber.getContactInfo().getId().equals(student.getContactInfo().getId())) {
-      return Response.status(Status.NOT_FOUND).build();
-    }
-
-    commonController.deletePhoneNumber(phoneNumber);
-
-    return Response.noContent().build();
   }
 
   @Path("/students/{STUDENTID:[0-9]*}/contactURLs")
